Use x.dtype for defualt h_0,c_0,bias, etc.
Given these should all be the same type T:

https://onnx.ai/onnx/operators/onnx__LSTM.html#l-onnx-doc-lstm
https://onnx.ai/onnx/operators/onnx__GRU.html#l-onnx-doc-gru
https://onnx.ai/onnx/operators/onnx__RNN.html#l-onnx-doc-rnn

+++ onnx/backend/test/case/node/gru.py
@@ -44,10 +44,17 @@ class GRUHelper:
 
         batch_size = x.shape[1]
         b = params.get(
-            B, np.zeros((self.num_directions, 2 * number_of_gates * hidden_size))
+            B,
+            np.zeros(
+                (self.num_directions, 2 * number_of_gates * hidden_size),
+                dtype=x.dtype,
+            ),
         )
         if h_0 is None:
-            h_0 = np.zeros((self.num_directions, batch_size, hidden_size))
+            h_0 = np.zeros(
+                (self.num_directions, batch_size, hidden_size),
+                dtype=x.dtype,
+            )
         lbr = params.get(LBR, 0)
 
         self.X = x
